requirements.txt
# selenium==4.27.1
# numpy==2.1.3
tqdm
fastapi
uvicorn
requests
pandas
google-cloud
google-cloud-bigquery
pyyaml
unidecode
python-jose
passlib[bcrypt]
pydantic
google-cloud-secret-manager